requirements.txt
gunicorn
django
django_heroku
django-lockdown
